Refactor the registration of transitions
Instead of using this complex dsl, we can simply leverage the `when` statement
to assign a transition to an event. It reduces code complexity and improve
clarity.

Side effect do not need to be supported the way it was after all, they can be
a part of the view state instead.